+++ packages/e2e-tests/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Unreleased
 
+- kill children in `runCommand` when parent process exits
 - make `deployAndDeposit` reserve 1000 eth instead of 100 eth to deployer
 - make `deployAndDeposit` wait for one confirmation on each tx
 - get rid of `startAnvil` script`